Restructure auth package and add show/hide password toggle
