+++ backend/src/main/java/com/swappable/backend/auth/AuthController.java
@@ -3,6 +3,13 @@ package com.swappable.backend.auth;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import com.swappable.backend.auth.dto.AuthResponse;
+import com.swappable.backend.auth.dto.LoginRequest;
+import com.swappable.backend.auth.dto.RegisterRequest;
+import com.swappable.backend.auth.dto.ForgotPasswordRequest;
+import com.swappable.backend.auth.dto.ResetPasswordRequest;
+import com.swappable.backend.auth.security.JwtService;
+import com.swappable.backend.auth.passwordreset.PasswordResetService;
 
 @RestController
 @RequestMapping("/api/auth")
