+++ backend/src/main/java/com/swappable/backend/auth/AuthService.java
@@ -6,6 +6,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import com.swappable.backend.auth.dto.AuthResponse;
+import com.swappable.backend.auth.dto.LoginRequest;
+import com.swappable.backend.auth.dto.RegisterRequest;
+import com.swappable.backend.auth.security.JwtService;
 
 @Service
 public class AuthService {
