+++ backend/src/main/java/com/swappable/backend/auth/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.dto;
 
 public record AuthResponse(
         String token,
