+++ backend/src/main/java/com/swappable/backend/auth/dto/ForgotPasswordRequest.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
