+++ backend/src/main/java/com/swappable/backend/auth/dto/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.dto;
 
 public record RegisterRequest(
         String username,
