+++ backend/src/main/java/com/swappable/backend/auth/dto/ResetPasswordRequest.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
