+++ backend/src/main/java/com/swappable/backend/auth/passwordreset/EmailService.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.passwordreset;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
