+++ backend/src/main/java/com/swappable/backend/auth/passwordreset/PasswordResetTokenRepository.java
@@ -1,7 +1,9 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.passwordreset;
 
 import com.swappable.backend.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.swappable.backend.auth.passwordreset.PasswordResetToken;
+
 
 import java.util.Optional;
