+++ backend/src/main/java/com/swappable/backend/auth/security/JwtService.java
@@ -1,4 +1,4 @@
-package com.swappable.backend.auth;
+package com.swappable.backend.auth.security;
 
 import com.swappable.backend.user.User;
 import io.jsonwebtoken.Jwts;
